refine captured dataloader
